refactor(team): extract promptBuilder from launcher.go (C1)
Lifts the system-prompt assembly out of Launcher into a free-standing
promptBuilder type with closure-based snapshot accessors (no Launcher,
Broker, or tmux dependency). Launcher.buildPrompt becomes a thin wrapper
that builds a promptBuilder from launcher state and delegates.

Tests-first: 13 new tests in prompt_builder_test.go drive the new type
directly, including branches that were 0%-covered before (1:1 mode,
codingAgentSlugs gh-pr-create instruction, focus-mode delegation block,
prompt determinism across member input orderings). The existing
TestBuildPrompt* characterization tests continue to exercise the same
output via Launcher.buildPrompt.

Per-file coverage on prompt_builder.go: 99.4%. Package coverage moves
62.9% -> 63.3%. launcher.go shrinks 4998 -> 4697 lines (-301, -6%).

First slice of the launcher decomposition described in PLAN.md.
